Move fonts from body to head
